Lab5/requirements.txt
pycryptodome>=3.19.0




